+++ unitTests/src/testIus3DTransducerElement.c
@@ -7,10 +7,8 @@
 #include <unity_internals.h>
 #include <unity_fixture.h>
 
-#include <include/ius.h>
-#include <include/iusError.h>
-#include <include/iusTypes.h>
-#include "include/ius3DTransducerElementPrivate.h"
+#include <ius.h>
+#include <ius3DTransducerElementPrivate.h>
 
 TEST_GROUP(Ius3DTransducerElement);
 
@@ -27,7 +25,7 @@ TEST(Ius3DTransducerElement, testIus3DTransducerElementCreate)
 {
     const float transducerPitch = 0.000005f;
     const int numTransducerElements = 128;
-    iu3dp_t elemPos = ius3DPositionCreate((10 - numTransducerElements / 2)*transducerPitch, 0.0f, 0.0f);
+    iu3dp_t elemPos = ius3DPositionCreate((float)(10 - numTransducerElements / 2.0)*transducerPitch, 0.0f, 0.0f);
     iu3ds_t elemSize = ius3DSizeCreate(0.0001f,0.0001f,0.0001f);
     iu3da_t elemAngle = ius3DAngleCreate(0.0f,0.3f);
     iu3dte_t element = ius3DTransducerElementCreate(elemPos, elemAngle, elemSize);
@@ -50,7 +48,7 @@ TEST(Ius3DTransducerElement, testIus3DTransducerElementDelete)
 {
     const float transducerPitch = 0.000005f;
     const int numTransducerElements = 128;
-    iu3dp_t elemPos = ius3DPositionCreate((10 - numTransducerElements / 2)*transducerPitch, 0.0f, 0.0f);
+    iu3dp_t elemPos = ius3DPositionCreate((float)(10 - numTransducerElements / 2.0)*transducerPitch, 0.0f, 0.0f);
     iu3ds_t elemSize = ius3DSizeCreate(0.0001f,0.0001f,0.0001f);
     iu3da_t elemAngle = ius3DAngleCreate(0.0f,0.3f);
     iu3dte_t element = ius3DTransducerElementCreate(elemPos, elemAngle, elemSize);
@@ -72,8 +70,8 @@ TEST(Ius3DTransducerElement, testIus3DTransducerElementCompare)
     IUS_BOOL equal;
     const float transducerPitch = 0.000005f;
     const int numTransducerElements = 128;
-    iu3dp_t elemPos = ius3DPositionCreate((10 - numTransducerElements / 2)*transducerPitch, 0.0f, 0.0f);
-    iu3dp_t diffElemPos = ius3DPositionCreate((19 - numTransducerElements / 2)*transducerPitch, 0.1f, 0.0f);
+    iu3dp_t elemPos = ius3DPositionCreate((float)(10 - numTransducerElements / 2.0)*transducerPitch, 0.0f, 0.0f);
+    iu3dp_t diffElemPos = ius3DPositionCreate((float)(19 - numTransducerElements / 2.0)*transducerPitch, 0.1f, 0.0f);
 
     iu3ds_t elemSize = ius3DSizeCreate(0.0001f,0.0001f,0.0001f);
     iu3ds_t diffElemSize = ius3DSizeCreate(0.0002f,0.0001f,0.0001f);
@@ -131,7 +129,7 @@ TEST(Ius3DTransducerElement, testIus3DTransducerElementSetGet)
     IUS_BOOL equal;
     const float transducerPitch = 0.000005f;
     const int numTransducerElements = 128;
-    iu3dp_t elemPos = ius3DPositionCreate((10 - numTransducerElements / 2)*transducerPitch, 0.0f, 0.0f);
+    iu3dp_t elemPos = ius3DPositionCreate((float)(10 - numTransducerElements / 2.0)*transducerPitch, 0.0f, 0.0f);
     iu3ds_t elemSize = ius3DSizeCreate(0.0001f,0.0001f,0.0001f);
     iu3da_t elemAngle = ius3DAngleCreate(0.0f,0.3f);
 
@@ -164,7 +162,7 @@ TEST(Ius3DTransducerElement, testIus3DTransducerElementSerialization)
     //char *path = "/TransducerElement";
     const float transducerPitch = 0.000005f;
     const int numTransducerElements = 128;
-    iu3dp_t elemPos = ius3DPositionCreate((10 - numTransducerElements / 2)*transducerPitch, 0.0f, 0.0f);
+    iu3dp_t elemPos = ius3DPositionCreate((float)(10 - numTransducerElements / 2.0)*transducerPitch, 0.0f, 0.0f);
     iu3ds_t elemSize = ius3DSizeCreate(0.0001f,0.0001f,0.0001f);
     iu3da_t elemAngle = ius3DAngleCreate(0.0f,0.3f);
